feat: Start splitting test specifications
