+++ dd-trace-api/src/main/java/datadog/trace/api/ConfigDefaults.java
@@ -11,7 +11,6 @@ public final class ConfigDefaults {
   public static final int DEFAULT_AGENT_TIMEOUT = 10; // timeout in seconds
   public static final String DEFAULT_SERVICE_NAME = "unnamed-java-app";
   public static final String DEFAULT_SERVLET_ROOT_CONTEXT_SERVICE_NAME = "root-servlet";
-  static final float DEFAULT_ANALYTICS_SAMPLE_RATE = 1.0f;
   public static final boolean DEFAULT_ASYNC_PROPAGATING = true;
   public static final int DEFAULT_TRACE_X_DATADOG_TAGS_MAX_LENGTH = 512;
